refactor: consolidate seedRoomMeta script
